devtools/madb: extract the app ID/activity name from the gradle scripts.
This is the first pass of extracting the app ID and the activity name
from the gradle based Android projects. madb start command can now be
invoked with no arguments, in which case the extracted ids are used.

It internally runs a custom Gradle script to figure out the correct
ids, so it can take a few seconds to get the ids, which is not ideal.
The following CL will address this issue by caching the most recently
extracted ids, so that they can be reused most of the time.

Change-Id: If9bad9ce6385a1e50bfffc6c37ee7ecb0752fe2a