madb: add version number to the binary
This CL adds 'madb version' command, which would print the version
number of the binary.

If the madb binary is one of the official releases, it 'madb version'
will print out the official version number.

Otherwise, it will print out '<version>-develop', where the version
number is the most recent release version prior to the version of this
binary.

The MADB_VERSION file contains the version number, and should be
considered as the standard place to keep the current version number.
The version string value within the 'version.go' file is updated via
'go generate' command, similar to how the doc.go / embedded_gradle.go
files are generated. This work is done in 'scripts/update_version.go'.
The potential version mismatch between MADB_VERSION file and the value
in 'version.go' will be caught by the 'madb-go-generate' test.

The 'scripts/release.go' is a script that packages the new version of
madb binaries for linux and darwin platforms.

The process of releasing a new version would be roughly as following:
- update the MADB_VERSION number and commit the CL.
- 'go generate' to update the version number within version.go
- 'go run scripts/release.go' to build the binaries
- draft a new release on GitHub with the new version number. attach
  the binaries built from the previous step.
- give the version tag using git, and push the tag to origin.
  (the mirroring job will mirror the tag to GitHub within 5min.)
- publish the release on GitHub.

Change-Id: Ibc96f50aa2237e81aef8b1e26ad4cfc8e4ba68fd
ref: #25