feat(config): migrate old config files to the new format
In commit 178b658, the old config files, "nicknames" and "users", were
merged into a single "config" file. From this CL, when madb sees the
old config files, it migrates those files into the new "config" file
so that the user doesn't lose all the nicknames and user IDs when
upgrading madb to a new version.

Change-Id: I6f376986dd55211729d0a5eb70dc885d26a93d9d